Move string-related stuff to stringify.ts
